requirements.txt
alembic==1.5.8
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.17.62
cachetools==4.2.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography
css-inline==0.6.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-CORS==3.0.10
Flask-JWT-Extended==4.0.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Sendgrid==0.7.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.2
google==3.0.0
google-api-core==1.26.0
google-auth==1.27.0
google-cloud==0.34.0
google-cloud-pubsub==2.4.0
google-cloud-secret-manager==2.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.36.0
gunicorn==20.0.4
html2text==2020.1.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.17
MarkupSafe==1.1.1
marshmallow==3.10.0
minify-html==0.4.10
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
nameparser==1.0.6
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
PyMySQL==1.0.2
pyngrok==5.0.5
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==0.16.1
workdays==1.4.0
